// planning-idsdp-pps-frontend/src/app/main/modules/rdpp-rtapp-management/models/rdpp-financial-analysis.model.ts
export class RdppFinancialAnalysis {
    uuid: string;
    projectConceptUuid: string;
    projectConceptMasterId: number;
    weatherAppraisalStudy: string;
    financialProjectLifeTime: number;
    economicProjectLifeTime: number;
    financialDiscountFactor: number;
    economicDiscountFactor: number;
    financialNpv: number;
    economicNpv: number;
    financialBcr: number;
    economicBcr: number;
    financialIrr: number;
    economicIrr: number;
    mainFeaturesOfRevision: string;
    cumulativeExpenditure: string;

    financialAnalysisAttachmentId: any;
    economicAnalysisAttachmentId: any;
}